Break out into a separate file for field creation. Should be mostly self contained.
